move bounds definitions inside type module
