+++ src/dsl/mod.rs
@@ -1,4 +1,3 @@
 pub mod compiler;
-pub mod enums;
 pub mod object_types;
 pub mod schema;
