+++ src/dsl/object_types/bounds/mod.rs
@@ -3,6 +3,29 @@ mod normalization;
 mod validation;
 
 use crate::dsl::schema::DisplayInformation;
+use serde_derive::Serialize;
+
+#[derive(Clone, Copy, Debug, Serialize)]
+pub enum IntegerBound {
+    Inclusive(i64),
+    Exclusive(i64),
+}
+
+#[derive(Clone, Copy, Debug)]
+pub struct IntegerObjectBounds {
+    pub minimum: Option<IntegerBound>,
+    pub maximum: Option<IntegerBound>,
+    pub multiple_of: Option<i64>,
+}
+
+impl IntegerBound {
+    pub fn inner(&self) -> &i64 {
+        match self {
+            IntegerBound::Exclusive(value) => value,
+            IntegerBound::Inclusive(value) => value,
+        }
+    }
+}
 
 #[derive(Clone, Debug)]
 pub struct EnumerationValues {
