+++ src/dsl/object_types/bounds/validation.rs
@@ -1,6 +1,6 @@
 use crate::dsl::compiler::validator::Validate;
 use crate::dsl::compiler::validator::ValidationError;
-use crate::dsl::enums::EnumerationValue;
+use crate::dsl::object_types::bounds::EnumerationValue;
 
 impl Validate for EnumerationValue {
     fn validate(&self) -> Result<(), ValidationError> {
